Minimize header includes by moving impl to .cpp files (#331)
* Minimize header includes by moving impl to .cpp files
* Make sure to build a shared library

(cherry picked from commit 1af8c802536537b0f383b36f5f1a89f708227d5a)

# Conflicts:
#	diagnostic_updater/include/diagnostic_updater/diagnostic_updater.hpp
