+++ diagnostic_updater/src/example.cpp
@@ -34,8 +34,9 @@
 
 #include <diagnostic_updater/diagnostic_updater.hpp>
 #include <diagnostic_updater/publisher.hpp>
-
 #include <std_msgs/msg/bool.hpp>
+
+#include "rclcpp/rclcpp.hpp"
 
 using namespace std::chrono_literals;
